Добавили изменения и доработали пару функции

+++ src/service/data-service/comment.js
@@ -6,25 +6,47 @@ class CommentService {
     this._Comment = sequelize.models.Comment;
   }
 
-  create(articleId, comment) {
-    return this._Comment.create({
+  async create(articleId, comment) {
+    return await this._Comment.create({
       articleId,
       ...comment
     });
   }
 
-  async drop(id) {
+  async drop(userId, articleId, commentId) {
+    const articleByUser = await this._Article.findOne({
+      where: {
+        id: articleId,
+        userId
+      }
+    });
+
+    if (!articleByUser) {
+      return !!articleByUser;
+    }
+
     const deletedRows = await this._Comment.destroy({
-      where: {id}
+      where: {
+        id: commentId
+      }
     });
 
     return !!deletedRows;
   }
 
-  findAll(articleId) {
-    return this._Comment.findAll({
+  async findAll(articleId) {
+    return await this._Comment.findAll({
       where: {articleId},
       raw: true
+    });
+  }
+
+  async findOne(id, articleId) {
+    return await this._Comment.findOne({
+      where: {
+        id,
+        articleId
+      }
     });
   }
 }
